+++ CHANGELOG.md
@@ -1,8 +1,15 @@
-### 1.2.0 - 2015-12-11
+Java Native Extractor by Fizzed
+===============================
+
+#### 2.0.0 - 2015-12-18
+ - Refactored package to com.fizzed
+ - Added slf4j for logging
+
+#### 1.2.0 - 2015-12-11
  - Refactored project layout
 
-### 1.1.0 - 2014-04-16
+#### 1.1.0 - 2014-04-16
  - Added ability to rename extracted file (e.g. cat to mycat)
 
-### 1.0.1 - 2014-03-18
+#### 1.0.1 - 2014-03-18
  - Initial release